requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.3.1
asn1crypto==1.5.1
async-timeout==5.0.1
attrs==24.2.0
bsv-sdk==0.5.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
coincurve==20.0.0
frozenlist==1.5.0
idna==3.10
multidict==6.1.0
propcache==0.2.1
pycparser==2.22
pycryptodomex==3.21.0
requests==2.32.3
typing_extensions==4.12.2
urllib3==2.2.3
yarl==1.18.3
